Moved FFmpeg RIFF utils to their own file
I'm going to add more later so it was getting kind of unwieldy
