+++ src/Makefile
@@ -25,7 +25,8 @@ CODING_OBJS=coding/adx_decoder.o \
     coding/at3_decoder.o \
     coding/g719_decoder.o \
     coding/hca_decoder.o \
-    coding/ffmpeg_decoder.o
+    coding/ffmpeg_decoder.o \
+    coding/ffmpeg_decoder_utils.o
 
 LAYOUT_OBJS=layout/ast_blocked.o \
     layout/blocked.o \
